pyproject.toml
[project]
name = "tabasco-fastapi"
version = "0.1.0"
description = "A FastAPI REST API for detecting intra-domain ambiguities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "kneed>=0.8.5",
    "nltk>=3.9.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "transformers>=4.52.4",
    "uv>=0.7.9",
]
